@@ -1,6 +1,6 @@
 {
     "name": "artplayer-packages",
-    "version": "5.1.0",
+    "version": "5.1.1",
     "type": "module",
     "private": true,
     "description": "ArtPlayer.js is a modern and full featured HTML5 video player",
@@ -49,14 +49,17 @@
         "cross-env": "^7.0.3",
         "cross-spawn": "^7.0.3",
         "dotenv": "^16.4.5",
-        "eslint": "^9.4.0",
-        "glob": "^10.4.1",
-        "globals": "^15.3.0",
-        "lerna": "^8.1.3",
-        "prettier": "^3.3.0",
+        "eslint": "^9.6.0",
+        "glob": "^11.0.0",
+        "globals": "^15.8.0",
+        "lerna": "^8.1.5",
+        "prettier": "^3.3.2",
         "prompts": "^2.4.2",
         "servor": "^4.0.2",
         "workbox-build": "^7.1.1"
+    },
+    "resolutions": {
+        "string-width": "4.2.3"
     },
     "engines": {
         "node": ">= 20.0.0"
