@@ -186,15 +186,19 @@ export default function artplayerPluginChapter(option = {}) {
 
 if (typeof document !== 'undefined') {
     const id = 'artplayer-plugin-chapter';
-    const $style = document.getElementById(id);
-    if ($style) {
-        $style.textContent = style;
-    } else {
-        const $style = document.createElement('style');
+    let $style = document.getElementById(id);
+    if (!$style) {
+        $style = document.createElement('style');
         $style.id = id;
-        $style.textContent = style;
-        document.head.appendChild($style);
+        if (document.readyState === 'loading') {
+            document.addEventListener('DOMContentLoaded', () => {
+                document.head.appendChild($style);
+            });
+        } else {
+            (document.head || document.documentElement).appendChild($style);
+        }
     }
+    $style.textContent = style;
 }
 
 if (typeof window !== 'undefined') {
