@@ -79,8 +79,8 @@ export default class Contextmenu extends Component {
         }
 
         proxy($player, 'contextmenu', (event) => {
-            event.preventDefault();
             if (!this.art.constructor.CONTEXTMENU) return;
+            event.preventDefault();
 
             this.show = true;
 
