@@ -8,6 +8,8 @@ export default function seekMix(art) {
     });
 
     def(art, 'loadedTime', {
-        get: () => ($video.buffered.length ? $video.buffered.end($video.buffered.length - 1) : 0),
+        get: () => {
+            return $video.buffered.length ? $video.buffered.end($video.buffered.length - 1) : 0;
+        },
     });
 }
