@@ -1,4 +1,4 @@
-import { inverseClass, queryAll, capitalize } from '../utils';
+import { capitalize } from '../utils';
 
 export default function flip(art) {
     const {
@@ -7,35 +7,36 @@ export default function flip(art) {
         constructor: { SETTING_ITEM_WIDTH, FLIP },
     } = art;
 
-    function update($panel, $tooltip, value) {
-        if ($tooltip) $tooltip.innerText = i18n.get(capitalize(value));
-        const $current = queryAll('.art-setting-item', $panel).find((item) => item.dataset.value === value);
-        if ($current) inverseClass($current, 'art-current');
+    function getI18n(value) {
+        return i18n.get(capitalize(value));
+    }
+
+    function update() {
+        const target = art.setting.find(`flip-${art.flip}`);
+        art.setting.check(target);
     }
 
     return {
         width: SETTING_ITEM_WIDTH,
         name: 'flip',
         html: i18n.get('Video Flip'),
-        tooltip: i18n.get(capitalize(art.flip)),
+        tooltip: getI18n(art.flip),
         icon: icons.flip,
         selector: FLIP.map((item) => {
             return {
                 value: item,
-                name: `aspect-ratio-${item}`,
+                name: `flip-${item}`,
                 default: item === art.flip,
-                html: i18n.get(capitalize(item)),
+                html: getI18n(item),
             };
         }),
         onSelect(item) {
             art.flip = item.value;
             return item.html;
         },
-        mounted: ($panel, item) => {
-            update($panel, item.$tooltip, art.flip);
-            art.on('flip', () => {
-                update($panel, item.$tooltip, art.flip);
-            });
+        mounted: () => {
+            update();
+            art.on('flip', () => update());
         },
     };
 }
