@@ -18,7 +18,8 @@ export default class Subtitle extends Component {
     constructor(art) {
         super(art);
         this.name = 'subtitle';
-        this.eventDestroy = () => null;
+        this.option = null;
+        this.destroyEvent = () => null;
         this.init(art.option.subtitle);
 
         let lastState = false;
@@ -42,11 +43,17 @@ export default class Subtitle extends Component {
     }
 
     get textTrack() {
-        return this.art.template.$video.textTracks[0];
+        return this.art.template.$video?.textTracks?.[0];
     }
 
-    get activeCue() {
-        return this.textTrack.activeCues[0];
+    get activeCues() {
+        if (!this.textTrack) return [];
+        return Array.from(this.textTrack.activeCues);
+    }
+
+    get cues() {
+        if (!this.textTrack) return [];
+        return Array.from(this.textTrack.cues);
     }
 
     style(key, value) {
@@ -58,19 +65,30 @@ export default class Subtitle extends Component {
     }
 
     update() {
-        const { $subtitle } = this.art.template;
+        const {
+            option: { subtitle },
+            template: { $subtitle },
+        } = this.art;
+
         $subtitle.innerHTML = '';
-        if (this.activeCue) {
-            if (this.art.option.subtitle.escape) {
-                $subtitle.innerHTML = this.activeCue.text
+        if (!this.activeCues.length) return;
+
+        this.art.emit('subtitleBeforeUpdate', this.activeCues);
+        $subtitle.innerHTML = this.activeCues
+            .map((cue, index) =>
+                cue.text
                     .split(/\r?\n/)
-                    .map((item) => `<div class="art-subtitle-line">${escape(item)}</div>`)
-                    .join('');
-            } else {
-                $subtitle.innerHTML = this.activeCue.text;
-            }
-            this.art.emit('subtitleUpdate', this.activeCue.text);
-        }
+                    .filter((line) => line.trim())
+                    .map(
+                        (line) =>
+                            `<div class="art-subtitle-line" data-group="${index}">
+                                ${subtitle.escape ? escape(line) : line}
+                            </div>`,
+                    )
+                    .join(''),
+            )
+            .join('');
+        this.art.emit('subtitleAfterUpdate', this.activeCues);
     }
 
     async switch(url, newOption = {}) {
@@ -93,13 +111,16 @@ export default class Subtitle extends Component {
         $newTrack.src = url;
         $newTrack.label = option.subtitle.name || 'Artplayer';
         $newTrack.track.mode = 'hidden';
+        $newTrack.onload = () => {
+            this.art.emit('subtitleLoad', this.cues, this.option);
+        };
 
-        this.eventDestroy();
+        this.art.events.remove(this.destroyEvent);
+        $track.onload = null;
         remove($track);
-
         append($video, $newTrack);
         template.$track = $newTrack;
-        this.eventDestroy = proxy(this.textTrack, 'cuechange', () => this.update());
+        this.destroyEvent = proxy(this.textTrack, 'cuechange', () => this.update());
     }
 
     async init(subtitleOption) {
@@ -108,8 +129,10 @@ export default class Subtitle extends Component {
             template: { $subtitle },
         } = this.art;
 
+        if (!this.textTrack) return null;
         validator(subtitleOption, scheme.subtitle);
         if (!subtitleOption.url) return;
+        this.option = subtitleOption;
         this.style(subtitleOption.style);
 
         return fetch(subtitleOption.url)
@@ -117,7 +140,6 @@ export default class Subtitle extends Component {
             .then((buffer) => {
                 const decoder = new TextDecoder(subtitleOption.encoding);
                 const text = decoder.decode(buffer);
-                this.art.emit('subtitleLoad', subtitleOption.url);
                 switch (subtitleOption.type || getExt(subtitleOption.url)) {
                     case 'srt': {
                         const vtt = srtToVtt(text);
@@ -142,7 +164,6 @@ export default class Subtitle extends Component {
                 if (this.url === subUrl) return subUrl;
                 URL.revokeObjectURL(this.url);
                 this.createTrack('metadata', subUrl);
-                this.art.emit('subtitleSwitch', subUrl);
                 return subUrl;
             })
             .catch((err) => {
