@@ -97,15 +97,19 @@ export function getIcon(key = '', html = '') {
 }
 
 export function setStyleText(id, style) {
-    const $style = document.getElementById(id);
-    if ($style) {
-        $style.textContent = style;
-    } else {
-        const $style = createElement('style');
+    let $style = document.getElementById(id);
+    if (!$style) {
+        $style = document.createElement('style');
         $style.id = id;
-        $style.textContent = style;
-        document.head.appendChild($style);
+        if (document.readyState === 'loading') {
+            document.addEventListener('DOMContentLoaded', () => {
+                document.head.appendChild($style);
+            });
+        } else {
+            (document.head || document.documentElement).appendChild($style);
+        }
     }
+    $style.textContent = style;
 }
 
 export function supportsFlex() {
@@ -116,4 +120,44 @@ export function supportsFlex() {
 
 export function getRect(el) {
     return el.getBoundingClientRect();
+}
+
+export function loadImg(url, scale) {
+    return new Promise((resolve, reject) => {
+        const img = new Image();
+
+        img.onload = function () {
+            if (!scale || scale === 1) {
+                resolve(img);
+            } else {
+                const canvas = document.createElement('canvas');
+                const ctx = canvas.getContext('2d');
+                canvas.width = img.width * scale;
+                canvas.height = img.height * scale;
+                ctx.drawImage(img, 0, 0, canvas.width, canvas.height);
+
+                canvas.toBlob((blob) => {
+                    const blobUrl = URL.createObjectURL(blob);
+                    const scaledImg = new Image();
+
+                    scaledImg.onload = function () {
+                        resolve(scaledImg);
+                    };
+
+                    scaledImg.onerror = function () {
+                        URL.revokeObjectURL(blobUrl);
+                        reject(new Error(`Image load failed: ${url}`));
+                    };
+
+                    scaledImg.src = blobUrl;
+                });
+            }
+        };
+
+        img.onerror = function () {
+            reject(new Error(`Image load failed: ${url}`));
+        };
+
+        img.src = url;
+    });
 }
