@@ -49,9 +49,9 @@ export type Utils = {
     secondToTime(second: number): string;
     escape(str: string): string;
     capitalize(str: string): string;
-    isStringOrNumber(val: any): boolean;
     getIcon(key: string, html: string | HTMLElement): HTMLElement;
     supportsFlex(): boolean;
     setStyleText(element: HTMLElement, text: string): void;
     getRect(el: HTMLElement): { top: number; left: number; width: number; height: number };
+    loadImg(url: string, scale?: Number): Promise<HTMLImageElement>;
 };
